Refactor match prediction daemon into standalone script
